requirements.txt
asgiref==3.4.1
attrs==22.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==3.2.9
django-cors-headers==3.10.0
django-debug-toolbar==3.2.4
django-extensions==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-model-utils==4.3.1
django-filter==23.2
drf-spectacular==0.21.0
drf-spectacular-sidecar==2021.11.29
factory-boy==3.2.0
Faker==30.0.0
gevent>=22.10.0
Cython>=0.29.30,<3.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==1.1.3.post0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
jsonschema==4.2.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==5.3.0
zope.event==5.0
zope.interface==7.0.3
